Women in Sports- Role Models
For my Text Analytics course, I analyzed the articles on women role models in sports. I wrote a report as if I was advising an athlete on what it takes to be a role model based on articles written on people's role models. I then verified this data by running an analysis on tweets about women role models in sports. 

Skills/methods I used:
- tokenization
- frequency counts
- word clouds
- sentiment analysis
- bar charts/ visualizations
- bigrams

In this repository there is:
- pdf submission of my final submission (a paper) with all code and output in the appendix
- word document with the articles copied and pasted into it